Move processes tests to own file
